✨: defaultVariants 주석처리
